API

Create netcdf file

NetcdfIO.create_nc! — Function

This function creates a netcdf file, and the supported methods are

create_nc!(file)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/create.jl:40.

create_nc!(file, dim_names, dim_sizes)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/create.jl:81.

NetcdfIO.create_nc! — Method

create_nc!(file::String)

Create an empty netcdf file, given

file Path to save the netcdf dataset

Examples

create_nc!("test.nc");

NetcdfIO.create_nc! — Method

create_nc!(file::String, dim_names::Vector{String}, dim_sizes::Vector)

Create an empty netcdf file with dimensions, given

file Path to save the netcdf dataset

dim_names Dimension names in the netcdf file

dim_sizes Sizes of the dimensions (must be Integer or Inf), the dimension is growable if size is Integer 0

Examples

create_nc!("test1.nc", String["lon", "lat", "ind"], [36, 18, 0]);
create_nc!("test2.nc", String["lon", "lat", "ind"], [36, 18, Inf]);

NetcdfIO.add_nc_dim! — Function

This function adds dim name and size information to netcdf file, and the supported methods are

add_nc_dim!(ds, dim_name, dim_size)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/create.jl:142.

add_nc_dim!(ds, dim_name, dim_size)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/create.jl:183.

add_nc_dim!(file, dim_name, dim_size)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/create.jl:215.

NetcdfIO.add_nc_dim! — Method

add_nc_dim!(ds::Dataset, dim_name::String, dim_size::Int)

Add dimension information to netcdf dataset, given

ds A NCDatasets.Dataset type dataset

dim_name Dimension name

dim_size Integer dimension size (0 for Inf, growable)

Examples

ds = Dataset("test.nc", "a");
add_nc_dim!(ds, "lat", 180);
add_nc_dim!(ds, "ind", 0);
close(ds);

NetcdfIO.add_nc_dim! — Method

add_nc_dim!(ds::Dataset, dim_name::String, dim_size::AbstractFloat)

Add dimension information to netcdf dataset, given

ds A NCDatasets.Dataset type dataset

dim_name Dimension name

dim_size Float dimension size, can be Inf

Examples

ds = Dataset("test.nc", "a");
add_nc_dim!("test.nc", "ind", Inf);
close(ds);

NetcdfIO.add_nc_dim! — Method

add_nc_dim!(file::String, dim_name::String, dim_size::Union{Int, AbstractFloat})

Add dimension information to netcdf file, given

file Path of the netcdf dataset

dim_name Dimension name

dim_size Dimension size, must be Inf or Integer

Examples

add_nc_dim!("test.nc", "lat", 180);
add_nc_dim!("test.nc", "ind", Inf);

Append new variables

NetcdfIO.append_nc! — Function

NCDatasets.jl does not have a convenient function (1 line command) to append or grow dataset into a file. Thus, we provide a few methods as supplements:

append_nc!(ds, var_name, var_data, var_attributes, dim_names; compress)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/append.jl:53.

append_nc!(file, var_name, var_data, var_attributes, dim_names; compress)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/append.jl:98.

NetcdfIO.append_nc! — Method

append_nc!(ds::Dataset, var_name::String, var_data::Array{T,N}, var_attributes::Dict{String,String}, dim_names::Vector{String}; compress::Int = 4) where {T<:Union{AbstractFloat,Int,String},N}

Append data to existing netcdf dataset, given

ds A NCDatasets.Dataset type dataset

var_name New variable name to write to

var_data New variable data to write, can be integer, float, and string with N dimens

var_attributes New variable attributes

dim_names Dimension names in the netcdf file

compress Compression level fro NetCDF, default is 4

Examples

create_nc!("test.nc", String["lon", "lat", "ind"], [36, 18, 5]);
dset = Dataset("test.nc", "a");
append_nc!(dset, "str", ["A" for i in 1:18], Dict("longname" => "test strings"), ["lat"]);
append_nc!(dset, "lat", collect(1:18), Dict("longname" => "latitude"), ["lat"]);
append_nc!(dset, "lon", collect(1:36), Dict("longname" => "longitude"), ["lon"]; compress=4);
append_nc!(dset, "ind", collect(1:5), Dict("longname" => "index"), ["ind"]);
append_nc!(dset, "d2d", rand(36,18), Dict("longname" => "a 2d dataset"), ["lon", "lat"]);
append_nc!(dset, "d3d", rand(36,18,5), Dict("longname" => "a 3d dataset"), ["lon", "lat", "ind"]);
close(dset);

NetcdfIO.append_nc! — Method

append_nc!(file::String, var_name::String, var_data::Array{T,N}, var_attributes::Dict{String,String}, dim_names::Vector{String}; compress::Int = 4) where {T<:Union{AbstractFloat,Int,String},N}

Append data to existing netcdf dataset, given

file Path of the netcdf dataset

var_name New variable name to write to

var_data New variable data to write, can be integer, float, and string with N dimens

var_attributes New variable attributes

dim_names Dimension names in the netcdf file

compress Compression level fro NetCDF, default is 4

Examples

create_nc!("test.nc", String["lon", "lat", "ind"], [36, 18, 5]);
append_nc!("test.nc", "str", ["A" for i in 1:18], Dict("longname" => "test strings"), ["lat"]);
append_nc!("test.nc", "lat", collect(1:18), Dict("longname" => "latitude"), ["lat"]);
append_nc!("test.nc", "lon", collect(1:36), Dict("longname" => "longitude"), ["lon"]; compress=4);
append_nc!("test.nc", "ind", collect(1:5), Dict("longname" => "index"), ["ind"]);
append_nc!("test.nc", "d2d", rand(36,18), Dict("longname" => "a 2d dataset"), ["lon", "lat"]);
append_nc!("test.nc", "d3d", rand(36,18,5), Dict("longname" => "a 3d dataset"), ["lon", "lat", "ind"]);

Grow existing variables

NetcdfIO.grow_nc! — Function

This function grows variables within a netcdf file, and the supported methods are

grow_nc!(ds, var_name, in_data, pending)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/grow.jl:55.

grow_nc!(file, var_name, in_data, pending)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/grow.jl:132.

NetcdfIO.grow_nc! — Method

grow_nc!(ds::Dataset, var_name::String, in_data::Union{AbstractFloat,Array,Int,String}, pending::Bool)

Grow the netcdf dataset, given

ds A NCDatasets.Dataset type dataset

var_name New variable name to write to

in_data New data to grow, can be integer, float, and string with N dimens

pending If true, the new data is appened to the end (growth); if false, the data will replace the ones from the bottom (when dimension has already growed)

Note that if there are more variables to grow at the same time, set pending to true only for the first time you call this function, and set pending to false for the rest variables.

Examples

create_nc!("test.nc", String["lon", "lat", "ind"], [36, 18, 0]);
dset = Dataset("test.nc", "a");
append_nc!(dset, "lat", collect(1:18), Dict("longname" => "latitude"), ["lat"]);
append_nc!(dset, "lon", collect(1:36), Dict("longname" => "longitude"), ["lon"]; compress=4);
append_nc!(dset, "ind", collect(1:5), Dict("longname" => "index"), ["ind"]);
append_nc!(dset, "d2d", rand(36,18), Dict("longname" => "a 2d dataset"), ["lon", "lat"]);
append_nc!(dset, "d3d", rand(36,18,5), Dict("longname" => "a 3d dataset"), ["lon", "lat", "ind"]);
grow_nc!(dset, "ind", 6, true);
grow_nc!(dset, "d3d", rand(36,18), false);
grow_nc!(dset, "d3d", rand(36,18), true);
grow_nc!(dset, "ind", 7, false);
close(dset);

NetcdfIO.grow_nc! — Method

grow_nc!(file::String, var_name::String, in_data::Union{AbstractFloat,Array,Int,String}, pending::Bool)

Grow the netcdf dataset, given

file Path of the netcdf dataset

var_name New variable name to write to

in_data New data to grow, can be integer, float, and string with N dimens

pending If true, the new data is appened to the end (growth); if false, the data will replace the ones from the bottom (when dimension has already growed)

Note that if there are more variables to grow at the same time, set pending to true only for the first time you call this function, and set pending to false for the rest variables.

Examples

create_nc!("test.nc", String["lon", "lat", "ind"], [36, 18, 0]);
dset = Dataset("test.nc", "a");
append_nc!(dset, "lat", collect(1:18), Dict("longname" => "latitude"), ["lat"]);
append_nc!(dset, "lon", collect(1:36), Dict("longname" => "longitude"), ["lon"]; compress=4);
append_nc!(dset, "ind", collect(1:5), Dict("longname" => "index"), ["ind"]);
append_nc!(dset, "d2d", rand(36,18), Dict("longname" => "a 2d dataset"), ["lon", "lat"]);
append_nc!(dset, "d3d", rand(36,18,5), Dict("longname" => "a 3d dataset"), ["lon", "lat", "ind"]);
close(dset);
grow_nc!("test.nc", "ind", 6, true);
grow_nc!("test.nc", "d3d", rand(36,18), false);
grow_nc!("test.nc", "d3d", rand(36,18), true);
grow_nc!("test.nc", "ind", 7, false);

Information of the dataset

NetcdfIO.dimname_nc — Function

dimname_nc(file::String)

Return all the names of the dimensions, given

file Path of the netcdf dataset

Examples

dims = dimname_nc("test.nc");

NetcdfIO.varname_nc — Function

This function returns the variable names in the netcdf dataset, and the supported methids are

varname_nc(file)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/info.jl:69.

varname_nc(ds)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/info.jl:100.

NetcdfIO.varname_nc — Method

varname_nc(file::String)

Return all the names of the variables (excluding the dimensions), given

file Path of the netcdf dataset

Examples

vars = varname_nc("test.nc");

NetcdfIO.varname_nc — Method

varname_nc(ds::Dataset)

Return all the names of the variables (excluding the dimensions), given

ds NCDatasets.Dataset type dataset

Examples

dset = Dataset("test.nc");
vars = varname_nc(dset);
close(dset);

NetcdfIO.size_nc — Function

size_nc(file::String, var_name::String)

Return the dimensions and size of a NetCDF dataset, given

file Path of the netcdf dataset

var_name Variable name

Examples

ndims,sizes = size_nc("test.nc", "test");

Read existing variables

NetcdfIO.find_variable — Function

find_variable(ds::Dataset, var_name::String)

Return the path to dataset if it exists, given

ds NCDatasets.Dataset type dataset

var_name Variable to read

NetcdfIO.read_nc — Function

NCDatasets.jl and NetCDF.jl both provide function to read data out from NC dataset. However, while NetCDF.jl is more convenient to use (less lines of code to read data), NCDatasets.jl is better to read a subset from the dataset and is able to detect the scale factor and offset. Here, we used a wrapper function to read NC dataset using NCDatasets.jl:

read_nc(file, var_name; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:86.

read_nc(T, file, var_name; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:138.

read_nc(file, var_name, indz; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:176.

read_nc(T, file, var_name, indz; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:226.

read_nc(file, var_name, indx, indy; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:263.

read_nc(T, file, var_name, indx, indy; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:315.

read_nc(file, var_name, indx, indy, indz; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:351.

read_nc(T, file, var_name, indx, indy, indz; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:397.

read_nc(file)
read_nc(file, selections; transform)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:433.

NetcdfIO.read_nc — Method

When only file name and variable label are provided, read_nc function reads out all the data:

read_nc(file::String, var_name::String; transform::Bool = true)

Read data from NC file, given

file Path of the netcdf dataset

var_name Variable to read

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Note that the missing data will be labeled as NaN.

Examples

# read data labeled as test from test.nc
save_nc!("test.nc", "test", rand(36,18,12), Dict("description" => "Random randoms"));
data = read_nc("test.nc", "test");

NetcdfIO.read_nc — Method

If a float type is given, the data will be converted to T, namely the output will be an array of T type numbers:

read_nc(T, file::String, var_name::String; transform::Bool = true)

Read data from nc file, given

T Number type

file Path of the netcdf dataset

var_name Variable name

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Examples

# read data labeled as test from test.nc as Float32
save_nc!("test.nc", "test", rand(36,18,12), Dict("description" => "Random randoms"));
data = read_nc(Float32, "test.nc", "test");

NetcdfIO.read_nc — Method

In many cases, the NC dataset can be very huge, and reading all the data points into one array could be time and memory consuming. In this case, reading a subset of data would be the best option:

read_nc(file::String, var_name::String, indz::Int; transform::Bool = true)

Read a subset from nc file, given

file Path of the netcdf dataset

var_name Variable name

indz The 3rd index of subset data to read

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Note that the dataset must be a 3D array to use this method.

Examples

# read 1st layer data labeled as test from test.nc
save_nc!("test.nc", "test", rand(36,18,12), Dict("description" => "Random randoms"));
data = read_nc("test.nc", "test", 1);

NetcdfIO.read_nc — Method

Similarly, one may want to read the subset as a certain type using

read_nc(T, file::String, var_name::String, indz::Int; transform::Bool = true)

Read a subset from nc file, given

T Number type

file Path of the netcdf dataset

var_name Variable name

indz The 3rd index of subset data to read

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Examples

# read 1st layer data labeled as test from test.nc as Float32
save_nc!("test.nc", "test", rand(36,18,12), Dict("description" => "Random randoms"));
data = read_nc(Float32, "test.nc", "test", 1);

NetcdfIO.read_nc — Method

Another convenient wrapper is to read all the data for given index in x and y, for example, if one wants to read the time series of data at a given site:

read_nc(file::String, var_name::String, indx::Int, indy::Int; transform::Bool = true)

Read the time series of data for a site, given

file Path of the netcdf dataset

var_name Variable name

indx The 1st index of subset data to read, typically longitude

indy The 2nd index of subset data to read, typically latitude

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Examples

save_nc!("test1.nc", "test", rand(36,18), Dict("description" => "Random randoms"));
save_nc!("test2.nc", "test", rand(36,18,12), Dict("description" => "Random randoms"));
data1 = read_nc("test1.nc", "test", 1, 1);
data2 = read_nc("test2.nc", "test", 1, 1);

NetcdfIO.read_nc — Method

Similarly, one may want to read the subset as a certain type using

read_nc(T, file::String, var_name::String, indx::Int, indy::Int; transform::Bool = true)

Read the time series of data for a site, given

T Number type

file Path of the netcdf dataset

var_name Variable name

indx The 1st index of subset data to read, typically longitude

indy The 2nd index of subset data to read, typically latitude

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Examples

save_nc!("test1.nc", "test", rand(36,18), Dict("description" => "Random randoms"));
save_nc!("test2.nc", "test", rand(36,18,12), Dict("description" => "Random randoms"));
data1 = read_nc(Float32, "test1.nc", "test", 1, 1);
data2 = read_nc(Float32, "test2.nc", "test", 1, 1);

NetcdfIO.read_nc — Method

Another convenient wrapper is to read the data for given index in x, y, and z, for example, if one wants to read the time series of data at a given site:

read_nc(file::String, var_name::String, indx::Int, indy::Int, indz::Int; transform::Bool = true)

Read the time series of data for a site, given

file Path of the netcdf dataset

var_name Variable name

indx The 1st index of subset data to read, typically longitude

indy The 2nd index of subset data to read, typically latitude

indz The 3rd index of subset data to read, typically time

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Examples

save_nc!("test.nc", "test", rand(36,18,12), Dict("description" => "Random randoms"));
data = read_nc("test.nc", "test", 1, 1, 1);

NetcdfIO.read_nc — Method

Similarly, one may want to read the data as a certain type using

read_nc(T, file::String, var_name::String, indx::Int, indy::Int, indz::Int; transform::Bool = true)

Read the time series of data for a site, given

T Number type

file Path of the netcdf dataset

var_name Variable name

indx The 1st index of subset data to read, typically longitude

indy The 2nd index of subset data to read, typically latitude

indz The 3rd index of subset data to read, typically time

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Examples

save_nc!("test.nc", "test", rand(36,18,12), Dict("description" => "Random randoms"));
data = read_nc(Float32, "test.nc", "test", 1, 1, 1);

NetcdfIO.read_nc — Function

The method below reads all the 1D data (with the same length) into a DataFrame

read_nc(file::String, selections::Vector{String} = varname_nc(file); transform::Bool = true)

Read the selected variables from a netcdf file as a DataFrame, given

file Path of the netcdf dataset

selections Variables to read from the file

transform If true, transform the data using NCDatasets rules, otherwise read the raw data

Examples

df_raw = DataFrame();
df_raw[!,"A"] = rand(5);
df_raw[!,"B"] = rand(5);
df_raw[!,"C"] = rand(5);
save_nc!("test.nc", df_raw);
df_new = read_nc("test.nc");
df_new = read_nc("test.nc", ["A", "B"]);

Quick save

NetcdfIO.save_nc! — Function

NCDatasets.jl does not have a convenient function (1 line command) to save dataset as a file. Thus, we provide a few methods as supplements:

save_nc!(file, var_name, var_data, var_attribute; compress, growable)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/save.jl:46.

save_nc!(file, df, var_names, var_attributes; compress, growable)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/save.jl:122.

save_nc!(file, df; compress, growable)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/save.jl:173.

NetcdfIO.save_nc! — Method

To save the code and effort to redefine the common attributes like latitude, longitude, and cycle index, we provide a shortcut method that handles these within the function:

save_nc!(file::String, var_name::String, var_data::Array{T,N}, var_attribute::Dict{String,String}; compress::Int = 4, growable::Bool = false) where {T<:Union{AbstractFloat,Int,String},N}

Save the 1D, 2D, or 3D data as netcdf file, given

file Path to save the dataset

var_name Variable name for the data in the NC file

var_data Data to save

var_attribute Variable attributes for the data, such as unit and long name

compress Compression level fro NetCDF, default is 4

growable If true, make index growable, default is false

Examples

# generate data to write into NC file
data1 = rand(12) .+ 273.15;
data2 = rand(36,18) .+ 273.15;
data3 = rand(36,18,12) .+ 273.15;

# save data as NC files (2D and 3D)
save_nc!("data1.nc", "data1", data1, Dict("description" => "Random temperature", "unit" => "K"));
save_nc!("data2.nc", "data2", data2, Dict("description" => "Random temperature", "unit" => "K"));
save_nc!("data3.nc", "data3", data3, Dict("description" => "Random temperature", "unit" => "K"));

NetcdfIO.save_nc! — Method

This method saves DataFrame as a NetCDF file to save more space (compared to a CSV file).

save_nc!(file::String, df::DataFrame, var_names::Vector{String}, var_attributes::Vector{Dict{String,String}}; compress::Int = 4, growable::Bool = false)

Save DataFrame to NetCDF, given

file Path to save the data

df DataFrame to save

var_names The label of data in DataFrame to save

var_attributes Variable attributes for the data to save

compress Compression level fro NetCDF, default is 4

growable If true, make index growable, default is false

Examples

df = DataFrame();
df[!,"A"] = rand(5);
df[!,"B"] = rand(5);
df[!,"C"] = rand(5);
save_nc!("dataf.nc", df, ["A","B"], [Dict("A" => "Attribute A"), Dict("B" => "Attribute B")]);

NetcdfIO.save_nc! — Method

This method is a simplified version of the method above, namely when users do not want to define the attributes.

save_nc!(file::String, df::DataFrame; compress::Int = 4, growable::Bool = false)

Save DataFrame to NetCDF, given

file Path to save the data

df DataFrame to save

notes Global attributes (notes)

compress Compression level fro NetCDF, default is 4

growable If true, make index growable, default is false

Examples

df = DataFrame();
df[!,"A"] = rand(5);
df[!,"B"] = rand(5);
df[!,"C"] = rand(5);
save_nc!("test.nc", df);