libnetcdf functions

Dataset IO

NetcdfIO.nc_open — Function

Open an existing netcdf dataset (nc_open)

NetcdfIO.nc_close — Function

Close a netcdf dataset (nc_close)

NetcdfIO.nc_create — Function

Create a new netcdf dataset (nc_create)

NetcdfIO.nc_redef — Function

Put the dataset into define mode (nc_redef)

NetcdfIO.nc_enddef — Function

End define mode for the dataset (nc_enddef)

NetcdfIO.nc_inq_path — Function

Inquire the file path of a netcdf dataset (ncinqpath)

Attribute related

NetcdfIO.nc_get_att — Function

Read attribute data from a netCDF dataset

NetcdfIO.nc_inq_att — Function

Inquire attribute information from a netCDF dataset

NetcdfIO.nc_inq_attname — Function

Inquire attribute name from a netCDF dataset

NetcdfIO.nc_put_att — Function

Write attribute data to a netCDF dataset

NetcdfIO.nc_put_att_string — Function

Write string attribute data to a netCDF dataset (calling ncputatt_string)

Dimension related

NetcdfIO.nc_def_dim — Function

Define the dimension with a netcdf dataset

NetcdfIO.nc_inq_dimid — Function

Inquire the dimension ID of a netcdf dataset (ncinqdimid)

NetcdfIO.nc_inq_dimids — Function

Inquire the dimension IDs of a netcdf dataset (ncinqdimids)

NetcdfIO.nc_inq_dimlen — Function

Inquire the dimension length of a netcdf dataset (ncinqdimlen)

NetcdfIO.nc_inq_dimname — Function

Inquire the dimension name of a netcdf dataset (ncinqdimname)

NetcdfIO.nc_inq_ndims — Function

Inquire the number of dimensions in a netcdf dataset (ncinqndims)

NetcdfIO.nc_inq_unlimdims — Function

Inquire the unlimited dimension IDs of a netcdf dataset (ncinqunlimdims)

Group related

NetcdfIO.nc_inq_grp_ncid — Function

Inquire the group ncid of a netcdf dataset (ncinqgrp_ncid)

NetcdfIO.nc_inq_grpname — Function

Inquire the group names of a netcdf dataset (ncinqgrpname)

NetcdfIO.nc_inq_grps — Function

Inquire the group IDs of a netcdf dataset (ncinqgrps)

Variable related

NetcdfIO.VariableLength — Type

Julia representation of the C struct ncvlent

NetcdfIO.nc_def_vlen — Function

Define a variable-length type in a netcdf dataset (ncdefvlen)

NetcdfIO.nc_free_vlen — Function

Free an array of VariableLength

NetcdfIO.nc_def_var — Function

Define a variable in a netcdf dataset

NetcdfIO.nc_def_var_deflate — Function

Define variable compression settings in a netcdf dataset

NetcdfIO.nc_get_var! — Function

Read data from a NetCDF variable into a Julia array (calling ncgetvar)

NetcdfIO.nc_get_var1 — Function

Read a single value from a NetCDF variable (ncgetvar1)

NetcdfIO.nc_get_vara! — Function

Read data from a NetCDF variable into a Julia array (calling ncgetvara)

NetcdfIO.nc_get_vars! — Function

Read data from a NetCDF variable into a Julia array (calling ncgetvars)

NetcdfIO.nc_put_var — Function

Write data to a netcdf dataset (ncputvar)

NetcdfIO.nc_put_var1 — Function

Write data to a netcdf dataset (ncputvar1)

NetcdfIO.nc_put_vara — Function

Write data to a netcdf dataset (ncputvara)

NetcdfIO.nc_put_vars — Function

Write data to a netcdf dataset (ncputvars)

NetcdfIO.nc_inq_user_type — Function

Inquire the user-defined type information of a netcdf dataset (ncinquser_type)

NetcdfIO.nc_inq_varnatts — Function

Inquire the number of attributes of a netcdf variable (ncinqvarnatts)

NetcdfIO.nc_inq_vardimid — Function

Inquire the variable dimension IDs of a netcdf dataset (ncinqvardimid)

NetcdfIO.nc_inq_varndims — Function

Inquire the number of dimensions of a netcdf variable (ncinqvarndims)

NetcdfIO.nc_inq_varid — Function

Inquire the variable ID of a netcdf dataset (ncinqvarid)

NetcdfIO.nc_inq_varids — Function

Inquire the variable IDs of a netcdf dataset (ncinqvarids)

NetcdfIO.nc_inq_varname — Function

Inquire the variable name of a netcdf dataset (ncinqvarname)

NetcdfIO.nc_inq_vartype — Function

Inquire the variable type of a netcdf dataset (ncinqvartype)

Error handling

NetcdfIO.NetCDFError — Type

Struct for netcdf error

NetcdfIO.nc_strerror — Function

Get netcdf error string from error code

NetcdfIO.check_status! — Function

Check the status code from a libnetcdf function call