API

Read variable size from netcdf

NetcdfIO.size_nc — Function

size_nc(file::String, var::String)

Return the dimensions and size of a NetCDF dataset, given

file Dataset path

var Variable name

Examples

ndims,sizes = read_nc("test.nc", "test");

Read variable from netcdf

NetcdfIO.read_nc — Function

NCDatasets.jl and NetCDF.jl both provide function to read data out from NC dataset. However, while NetCDF.jl is more convenient to use (less lines of code to read data), NCDatasets.jl is better to read a subset from the dataset and is able to detect the scale factor and offset. Here, we used a wrapper function to read NC dataset using NCDatasets.jl:

read_nc(file, var)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:29.

read_nc(T, file, var)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:59.

read_nc(file, var, indz)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:81.

read_nc(T, file, var, indz)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:112.

read_nc(file, var, indx, indy)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:132.

read_nc(T, file, var, indx, indy)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:163.

read_nc(file, var, indx, indy, indz)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:184.

read_nc(T, file, var, indx, indy, indz)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/read.jl:212.

NetcdfIO.read_nc — Method

When only file name and variable label are provided, read_nc function reads out all the data:

read_nc(file::String, var::String)

Read data from NC file, given

file Dataset path

var Variable to read

Note that the missing data will be labeled as NaN.

Examples

# read data labeled as test from test.nc
data = read_nc("test.nc", "test");

NetcdfIO.read_nc — Method

If a float type is given, the data will be converted to T, namely the output will be an array of T type numbers:

read_nc(T, file::String, var::String)

Read data from nc file, given

T Number type

file Dataset path

var Variable name

Examples

# read data labeled as test from test.nc as Float32
data = read_nc(Float32, "test.nc", "test");

NetcdfIO.read_nc — Method

In many cases, the NC dataset can be very huge, and reading all the data points into one array could be time and memory consuming. In this case, reading a subset of data would be the best option:

read_nc(file::String, var::String, indz::Int)

Read a subset from nc file, given

file Dataset path

var Variable name

indz The 3rd index of subset data to read

Note that the dataset must be a 3D array to use this method.

Examples

# read 1st layer data labeled as test from test.nc
data = read_nc("test.nc", "test", 1);

NetcdfIO.read_nc — Method

Similarly, one may want to read the subset as a certain type using

read_nc(T, file::String, var::String, indz::Int)

Read a subset from nc file, given

T Number type

file Dataset path

var Variable name

indz The 3rd index of subset data to read

Examples

# read 1st layer data labeled as test from test.nc as Float32
data = read_nc(Float32, "test.nc", "test", 1);

NetcdfIO.read_nc — Method

Another convenient wrapper is to read all the data for given index in x and y, for example, if one wants to read the time series of data at a given site:

read_nc(file::String, var::String, indx::Int, indy::Int)

Read the time series of data for a site, given

file Dataset path

var Variable name

indx The 1st index of subset data to read, typically longitude

indy The 2nd index of subset data to read, typically latitude

Examples

data = read_nc("test.nc", "test", 1, 1);

NetcdfIO.read_nc — Method

Similarly, one may want to read the subset as a certain type using

read_nc(T, file::String, var::String, indx::Int, indy::Int)

Read the time series of data for a site, given

T Number type

file Dataset path

var Variable name

indx The 1st index of subset data to read, typically longitude

indy The 2nd index of subset data to read, typically latitude

Examples

data = read_nc(Float32, "test.nc", "test", 1, 1);

NetcdfIO.read_nc — Method

Another convenient wrapper is to read the data for given index in x, y, and z, for example, if one wants to read the time series of data at a given site:

read_nc(file::String, var::String, indx::Int, indy::Int, indz::Int)

Read the time series of data for a site, given

file Dataset path

var Variable name

indx The 1st index of subset data to read, typically longitude

indy The 2nd index of subset data to read, typically latitude

indz The 3rd index of subset data to read, typically time

Examples

data = read_nc("test.nc", "test", 1, 1, 1);

NetcdfIO.read_nc — Method

Similarly, one may want to read the data as a certain type using

read_nc(T, file::String, var::String, indx::Int, indy::Int, indz::Int)

Read the time series of data for a site, given

T Number type

file Dataset path

var Variable name

indx The 1st index of subset data to read, typically longitude

indy The 2nd index of subset data to read, typically latitude

indz The 3rd index of subset data to read, typically time

Examples

data = read_nc(Float32, "test.nc", "test", 1, 1, 1);

Save variable to netcdf

NetcdfIO.save_nc! — Function

NCDatasets.jl does not have a convenient function (1 line command) to save dataset as a file. Thus, we provide a few methods as supplements:

save_nc!(file, var_name, var_attr, var_data, atts_name, atts_attr, atts_data, notes; compress)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/save.jl:79.

save_nc!(file, var_name, var_attr, var_data; notes, compress)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/save.jl:153.

save_nc!(file, var_names, var_attrs, df; notes, compress)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/save.jl:216.

save_nc!(file, df; notes, compress)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/save.jl:250.

NetcdfIO.save_nc! — Method

This method is a case if one wants to save both variable and attributes into the target file. This method support saving multiple (N) dimension arrays:

save_nc!(file::String,
         var_name::String,
         var_attr::Dict{String,String},
         var_data::Array{T,N},
         atts_name::Vector{String},
         atts_attr::Vector{Dict{String,String}},
         atts_data::Vector{Vector},
         notes::Dict{String,String};
         compress::Int = 4
) where {T<:Union{AbstractFloat,Int,String},N}

Save dataset as NC file, given

file Path to save the dataset

var_name Variable name for the data in the NC file

var_attr Variable attributes for the data, such as unit and long name

var_data Data to save

atts_name vector of supporting attribute labels, such as lat and lon

atts_attr Vector of attributes for the supporting attributes, such as unit

atts_data Vector of attributes data, such as the latitude range

notes Global attributes (notes)

compress Compression level fro NetCDF, default is 4

Examples

# generate data to write into NC file
lats = collect(Float64, -85:10:85);
lons = collect(Float64, -175:10:175);
inds = collect(Int, 1:12);
data1 = rand(18) .+ 273.15;
data2 = rand(36,18) .+ 273.15;
data3 = rand(36,18,12) .+ 273.15;

# define the attributes of the dimensions and data
attrn = Dict("description" => "Random temperature", "unit" => "K");
latat = Dict("description" => "Latitude", "unit" => "°");
lonat = Dict("description" => "Longitude", "unit" => "°");
indat = Dict("description" => "Cycle index", "unit" => "-");

# define attributes names, information, and data
atts_name1 = ["lat"];
atts_name2 = ["lon", "lat"];
atts_name3 = ["lon", "lat", "ind"];
atts_attr1 = [latat];
atts_attr2 = [lonat, latat];
atts_attr3 = [lonat, latat, indat];
atts_data1 = Any[lats];
atts_data2 = Any[lons, lats];
atts_data3 = Any[lons, lats, inds];
notes = Dict("description" => "This is a file generated using PkgUtility.jl", "notes" => "PkgUtility.jl uses NCDatasets.jl to create NC files");

# save data as NC files (1D, 2D, and 3D)
save_nc!("data1.nc", "data1", attrn, data1, atts_name1, atts_attr1, atts_data1, notes);
save_nc!("data2.nc", "data2", attrn, data2, atts_name2, atts_attr2, atts_data2, notes);
save_nc!("data3.nc", "data3", attrn, data3, atts_name3, atts_attr3, atts_data3, notes);

NetcdfIO.save_nc! — Method

To save the code and effort to redefine the common attributes like latitude, longitude, and cycle index, we provide a shortcut method that handles these within the function:

save_nc!(file::String,
         var_name::String,
         var_attr::Dict{String,String},
         var_data::Array{T,N};
         notes::Dict{String,String} = ATTR_ABOUT,
         compress::Int = 4
) where {T<:Union{AbstractFloat,Int,String},N}

Save the 2D or 3D data as NC file, given

file Path to save the dataset

var_name Variable name for the data in the NC file

var_attr Variable attributes for the data, such as unit and long name

var_data Data to save

notes Global attributes (notes)

compress Compression level fro NetCDF, default is 4

Examples

# generate data to write into NC file
data2 = rand(36,18) .+ 273.15;
data3 = rand(36,18,12) .+ 273.15;

# define the attributes and notes
attrn = Dict("description" => "Random temperature", "unit" => "K");
notes = Dict("description" => "This is a file generated using PkgUtility.jl", "notes" => "PkgUtility.jl uses NCDatasets.jl to create NC files");

# save data as NC files (2D and 3D)
save_nc!("data2.nc", "data2", attrn, data2);
save_nc!("data2.nc", "data2", attrn, data2; notes=notes);
save_nc!("data3.nc", "data3", attrn, data3);
save_nc!("data3.nc", "data3", attrn, data3; notes=notes);

NetcdfIO.save_nc! — Method

This method saves DataFrame as a NetCDF file to save more space (compared to a CSV file).

save_nc!(file::String, var_names::Vector{String}, var_attrs::Vector{Dict{String,String}}, df::DataFrame; notes::Dict{String,String} = ATTR_ABOUT, compress::Int = 4)

Save DataFrame to NetCDF, given

file Path to save the data

var_names The label of data in DataFrame to save

var_attrs Variable attributes for the data to save

df DataFrame to save

notes Global attributes (notes)

compress Compression level fro NetCDF, default is 4

Examples

df = DataFrame();
df[!,"A"] = rand(5);
df[!,"B"] = rand(5);
df[!,"C"] = rand(5);
save_nc!("test.nc", ["A","B"], [Dict("A" => "Attribute A"), Dict("B" => "Attribute B")], df);

NetcdfIO.save_nc! — Method

This method is a simplified version of the method above, namely when users do not want to define the attributes.

save_nc!(file::String, df::DataFrame; notes::Dict{String,String} = ATTR_ABOUT, compress::Int = 4)

Save DataFrame to NetCDF, given

file Path to save the data

df DataFrame to save

notes Global attributes (notes)

compress Compression level fro NetCDF, default is 4

Examples

df = DataFrame();
df[!,"A"] = rand(5);
df[!,"B"] = rand(5);
df[!,"C"] = rand(5);
save_nc!("test.nc", df);

Append variable to netcdf

NetcdfIO.append_nc! — Function

NCDatasets.jl does not have a convenient function (1 line command) to append dataset into a file. Thus, we provide a few methods as supplements:

append_nc!(file, var_name, var_attr, var_data, atts_name, atts_attr, atts_data; compress)

defined at /home/runner/work/NetcdfIO.jl/NetcdfIO.jl/src/append.jl:76.

NetcdfIO.append_nc! — Method

This method append data to an exisiting NC file. If the attributes exist already, then only save the data:

append_nc!(file::String,
           var_name::String,
           var_attr::Dict{String,String},
           var_data::Array{FT,N},
           atts_name::Vector{String},
           atts_attr::Vector{Dict{String,String}},
           atts_data::Vector;
           compress::Int = 4
) where {FT<:AbstractFloat,N}

Append data to existing file, given

file Path to save the dataset

var_name Variable name for the data in the NC file

var_attr Variable attributes for the data, such as unit and long name

var_data Data to save

atts_name vector of supporting attribute labels, such as lat and lon

atts_attr Vector of attributes for the supporting attributes, such as unit

atts_data Vector of attributes data, such as the latitude range

compress Compression level fro NetCDF, default is 4

Examples

# generate data to write into NC file
lats = collect(Float64, -85:10:85);
lons = collect(Float64, -175:10:175);
inds = collect(Int, 1:12);
data1 = rand(18) .+ 273.15;
data2 = rand(36,18) .+ 273.15;
data3 = rand(36,18,12) .+ 273.15;

# define the attributes of the dimensions and data
attrn = Dict("description" => "Random temperature", "unit" => "K");
latat = Dict("description" => "Latitude", "unit" => "°");
lonat = Dict("description" => "Longitude", "unit" => "°");
indat = Dict("description" => "Cycle index", "unit" => "-");

# define attributes names, information, and data
atts_name1 = ["lat"];
atts_name2 = ["lon", "lat"];
atts_name3 = ["lon", "lat", "ind"];
atts_attr1 = [latat];
atts_attr2 = [lonat, latat];
atts_attr3 = [lonat, latat, indat];
atts_data1 = Any[lats];
atts_data2 = Any[lons, lats];
atts_data3 = Any[lons, lats, inds];
notes = Dict("description" => "This is a file generated using PkgUtility.jl", "notes" => "PkgUtility.jl uses NCDatasets.jl to create NC files");

# save data as NC files (1D, 2D, and 3D)
append_nc!("data1.nc", "datax", attrn, data1, atts_name1, atts_attr1, atts_data1);
append_nc!("data2.nc", "datax", attrn, data2, atts_name2, atts_attr2, atts_data2);
append_nc!("data3.nc", "datax", attrn, data3, atts_name3, atts_attr3, atts_data3);